Fix IE, userfriendly, autosolve, directional lines
Fixed IE compatibility by removing 'for of' loop and 'output' tag.
When the reveal bar is at the maximum, it will stay at the maximum until
changed.
Added two optional features:
Autosolve - Any changes to the board will attempt to be solved.
Directional lines - When turned on, lines are thicker in one direction,
to give a sense of which direction to go.