Adding directional arrows.
Thanks to @Nickardson in pull request #2 for the idea!

These arrows look good, but someday in the future, if the algorithm
changes, we might need to either remove the arrows or implement an
additional step to calculate the arrow direction. In other words, these
arrows must not hinder future changes to the algorithm.