Fixing `marker-mid` rendering.
Turns out Chrome rendering is incorrect, and Firefox was correct.

Relative URIs in external CSS files are resolved using the stylesheet
file as base, and not the document. This means `marker-mid:
url(#foobar)` should not work in an external CSS file. Google Chrome
renders the arrows, Firefox doesn't. And Firefox is correct.

* https://bugzilla.mozilla.org/show_bug.cgi?id=309612#c5
* http://codepen.io/denilsonsa/pen/yNpXVK

The solution implemented here is to specify `marker-mid` in the
generated SVG file, and override it to `none` via CSS. This should work
in every browser.